fix(day-3): covert adjecent number postions to a set and fix bug introduced by refactor
Using a set instead of vector to hold position of adjecent numbers in the event two symbols are adjecent to the same digit. Also, fixed a bug discovered when printing the length of the resultant set introduced from a previous refactor that was causing the build of the set to happen inside a loop instead of just once.