refactor(day-3): rename structs and associated files
As loop for building the number string from the digit position was being constructed decided to rename structs (and associated files) to match the naming convention in the problem statement. Added convience method to SchematicObj enum to retrieve value. Moved the number string builder method to an associated function since it processes a single row and doesn't require the entire matrix.